feat: movie lists now shows up in the library screen.

+++ ui/src/main/java/com/cairosquad/ui/library/component/ListContainer.kt
@@ -32,7 +32,7 @@ import com.cairosquad.design_system.theme.Theme
 
 @Composable
 fun ListContainer(
-	numberOfItems: Int,
+	numberOfItems: Long,
 	listName: String,
 	onListClicked: () -> Unit,
 	modifier: Modifier = Modifier,
